requirements.txt
# Core Logic & Data
pandas
numpy

# API & Web Services
fastapi
uvicorn
aiofiles

# Exchange Integration
ccxt==4.*

# Database
sqlite-utils==3.*

# Testing
pytest
pytest-asyncio
pytest-mock
httpx 

# UI & Dashboard
streamlit
streamlit-autorefresh
python-dotenv